Fixes package import setup issue
