+++ i8_terminal/version.txt
@@ -1,1 +1,1 @@
-__version__ = "0.2.67"
+__version__ = "0.2.68"
